+++ packages/yourloops/components/statistics/glucose-management-indicator-stat.tsx
@@ -36,7 +36,7 @@ interface GlucoseManagementIndicatorProps {
 
 export const GlucoseManagementIndicator: FunctionComponent<GlucoseManagementIndicatorProps> = (props) => {
   const { glucoseManagementIndicator } = props
-  const annotations = glucoseManagementIndicator ? [t('glucose-management-indicator-tooltip')] : [t('glucose-management-indicator-tooltip'), t('glucose-management-indicator-empty-stat')]
+  const annotations = glucoseManagementIndicator ? [t('glucose-management-indicator-tooltip')] : [t('glucose-management-indicator-tooltip'), t('tooltip-empty-stat')]
 
   return (
     <Box data-testid="glucose-management-indicator-stat">